Reorganized event listeners (task #4340)
